# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /** Platform role — user or admin. */
  role: string;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type EventViewOutputSnapshot = { [key: string]: unknown };

export interface EventView {
  id: string;
  type: string;
  description: string;
  stage: string;
  created_at: string;
  output_snapshot?: EventViewOutputSnapshot;
}

export interface ContactRef {
  id: string;
}

export interface OpportunityView {
  opportunity_id: string;
  company: string;
  created_at: string;
  events: EventView[];
  contacts: ContactRef[];
}

export interface OpportunityListEnvelope {
  opportunities: OpportunityView[];
}

export interface CreateOpportunityRequest {
  /** @minLength 1 */
  company: string;
}

export interface UpdateOpportunityRequest {
  name?: string;
  customerName?: string;
  status?: string;
  stage?: string;
  summary?: string;
}

export type AddEventRequestOutputSnapshot = { [key: string]: unknown };

export interface AddEventRequest {
  type?: string;
  description?: string;
  stage?: string;
  output_snapshot?: AddEventRequestOutputSnapshot;
}

export interface ArchiveResult {
  archived: boolean;
  id: string;
}

/**
 * Arbitrary JSON payload representing the full LogicGate session state.
 */
export interface WorkingStatePayload { [key: string]: unknown }

export interface WorkingStateEnvelope {
  state: unknown | null;
}

export interface WorkingStateSaveResult {
  savedAt: string;
}

/**
 * Not authenticated.
 */
export type UnauthorisedResponse = ErrorEnvelope;

/**
 * Authenticated but not authorised.
 */
export type ForbiddenResponse = ErrorEnvelope;

/**
 * Resource not found.
 */
export type NotFoundResponse = ErrorEnvelope;

/**
 * Invalid request body.
 */
export type BadRequestResponse = ErrorEnvelope;

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};