# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /opportunities:
    get:
      tags: [Opportunities]
      operationId: listOpportunities
      summary: List active opportunities for the authenticated user's organisation
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Opportunity list returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Opportunities]
      operationId: createOpportunity
      summary: Create a new opportunity
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportunityRequest'
      responses:
        '201':
          description: Opportunity created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
  /opportunities/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    get:
      tags: [Opportunities]
      operationId: getOpportunity
      summary: Get a single opportunity with events and contact refs
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Opportunity found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityView'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Opportunities]
      operationId: updateOpportunity
      summary: Update an opportunity
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOpportunityRequest'
      responses:
        '200':
          description: Opportunity updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityView'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Opportunities]
      operationId: archiveOpportunity
      summary: Archive an opportunity
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Opportunity archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveResult'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /opportunities/{id}/events:
    post:
      tags: [Opportunities]
      operationId: addOpportunityEvent
      summary: Append an event to an opportunity
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/EntityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddEventRequest'
      responses:
        '201':
          description: Event appended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventView'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /opportunities/{id}/working-state:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    get:
      tags: [Opportunities]
      operationId: getWorkingState
      summary: Load the working session state for an opportunity
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Working state returned (state is null if none saved yet).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingStateEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Opportunities]
      operationId: saveWorkingState
      summary: Upsert the working session state for an opportunity
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkingStatePayload'
      responses:
        '200':
          description: State saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingStateSaveResult'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
    EntityId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          description: Platform role — user or admin.
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    EventView:
      type: object
      required: [id, type, description, stage, created_at]
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        stage:
          type: string
        created_at:
          type: string
          format: date-time
        output_snapshot:
          type: object
    ContactRef:
      type: object
      required: [id]
      properties:
        id:
          type: string
    OpportunityView:
      type: object
      required: [opportunity_id, company, created_at, events, contacts]
      properties:
        opportunity_id:
          type: string
        company:
          type: string
        created_at:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventView'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactRef'
    OpportunityListEnvelope:
      type: object
      required: [opportunities]
      properties:
        opportunities:
          type: array
          items:
            $ref: '#/components/schemas/OpportunityView'
    CreateOpportunityRequest:
      type: object
      required: [company]
      properties:
        company:
          type: string
          minLength: 1
    UpdateOpportunityRequest:
      type: object
      properties:
        name:
          type: string
        customerName:
          type: string
        status:
          type: string
        stage:
          type: string
        summary:
          type: string
    AddEventRequest:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        stage:
          type: string
        output_snapshot:
          type: object
    ArchiveResult:
      type: object
      required: [archived, id]
      properties:
        archived:
          type: boolean
        id:
          type: string
    WorkingStatePayload:
      type: object
      additionalProperties: true
      description: Arbitrary JSON payload representing the full LogicGate session state.
    WorkingStateEnvelope:
      type: object
      required: [state]
      properties:
        state:
          nullable: true
    WorkingStateSaveResult:
      type: object
      required: [savedAt]
      properties:
        savedAt:
          type: string
          format: date-time
  responses:
    Unauthorised:
      description: Not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but not authorised.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Invalid request body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'