# lib/api-zod/src/generated/types/addEventRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AddEventRequestOutputSnapshot } from './addEventRequestOutputSnapshot';

export interface AddEventRequest {
  type?: string;
  description?: string;
  stage?: string;
  output_snapshot?: AddEventRequestOutputSnapshot;
}