# lib/api-zod/src/generated/types/createOpportunityRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOpportunityRequest {
  /** @minLength 1 */
  company: string;
}