# lib/api-zod/src/generated/types/eventView.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventViewOutputSnapshot } from './eventViewOutputSnapshot';

export interface EventView {
  id: string;
  type: string;
  description: string;
  stage: string;
  created_at: Date;
  output_snapshot?: EventViewOutputSnapshot;
}