# lib/api-zod/src/generated/types/opportunityListEnvelope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityView } from './opportunityView';

export interface OpportunityListEnvelope {
  opportunities: OpportunityView[];
}