# lib/api-zod/src/generated/types/opportunityView.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactRef } from './contactRef';
import type { EventView } from './eventView';

export interface OpportunityView {
  opportunity_id: string;
  company: string;
  created_at: Date;
  events: EventView[];
  contacts: ContactRef[];
}