# lib/api-zod/src/generated/types/updateOpportunityRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOpportunityRequest {
  name?: string;
  customerName?: string;
  status?: string;
  stage?: string;
  summary?: string;
}