# lib/api-zod/src/generated/types/workingStatePayload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Arbitrary JSON payload representing the full LogicGate session state.
 */
export interface WorkingStatePayload { [key: string]: unknown }